+++ src/Logic/Disjunction.php
@@ -105,6 +105,11 @@ final class Disjunction implements Expression
         return $this->orX(Expr::key($key, $expr));
     }
 
+    public function orMethod($methodName, Expression $expr)
+    {
+        return $this->orX(Expr::method($methodName, $expr));
+    }
+
     public function orAtLeast($count, Expression $expr)
     {
         return $this->orX(Expr::atLeast($count, $expr));