+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+* 1.0.0-beta5 (@release_date@)
+
+ * added `method()` selector
+
 * 1.0.0-beta4 (2015-08-24)
 
  * fixed return types in `Expr::true()` and `Expr::false()`